requirements.txt
Flask==1.0.2
pytest==4.3.1
requests==2.21.0
StringGenerator==0.3.1
python-twitter==3.5
